directory correction
