+++ controllers/db_handler.go
@@ -1,0 +1,26 @@
+package controllers
+
+import (
+	"database/sql"
+	"log"
+
+	"gorm.io/driver/mysql"
+	"gorm.io/gorm"
+)
+
+func connect() *sql.DB {
+	db, err := sql.Open("mysql", "root:@tcp(localhost:3306)/musify?parseTime=true&loc=Asia%2FJakarta")
+	if err != nil {
+		log.Fatal(err)
+	}
+	return db
+}
+
+func connectGorm() *gorm.DB {
+	dsn := "root:@tcp(localhost:3306)/musify?parseTime=true&loc=Asia%2FJakarta"
+	db, err := gorm.Open(mysql.Open(dsn), &gorm.Config{})
+	if err != nil {
+		log.Fatal(err)
+	}
+	return db
+}
